adding exercises 4.22 and 4.23

+++ part4/blogs/controllers/blogs.js
@@ -11,12 +11,16 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
+
+  if (body.title === undefined || body.url === undefined) {
+    return response.status(400).json({ error: 'title and url are mandatory' })
   }
 
-  const user = await User.findById(decodedToken.id)
+  if (body.likes === undefined) {
+    body.likes = 0
+  }
+
+  const user = request.user
 
   const newBlog = new Blog({
     title: body.title,
@@ -34,6 +38,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const body = request.body
+
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (blogToDelete.user != request.user.id) {
+    console.log('BLOG USER', blogToDelete.user, 'REQUEST USER', request.user.id)
+    return response.status(401).json({ error: 'not your post. cannot delete.' })
+  }
+
+  const user = request.user
   const result = await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
